feat(type-infomation): type of start function now available
You can now query the types of the arguments of the start
function.

It is available at:

`plugin.typeEnvironment.startTypes`

The type of start types is `Type[][]`. Apparently typescript allows for
function overloading, so it is an array of array of argument lists.

examples can be found at test/test-plugin.ts

fix(validation): plugin now checks that a start function exists

BREAKING CHANGE: plugin constructor (and thus loadPlugin) now
throw an error if there is no start function, (this may be reverted).